% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-posthoc-quantities.R
\name{add_log_posterior_values}
\alias{add_log_posterior_values}
\title{Approximation to hyperparameter log-posterior}
\usage{
add_log_posterior_values(optresults, model_data)
}
\arguments{
\item{optresults}{Optimization results, a tibble() output by optimize_all_thetas_parallel().}

\item{model_data}{ccmodeldata object output by model_setup()}
}
\value{
A tibble() in the same format as optresults, with columns added for the log-posterior
of theta and sigma.
}
\description{
This function computes log pi(theta|y) for all thetas in the grid.
}
