% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseCrossover.R
\docType{package}
\name{CaseCrossover-package}
\alias{CaseCrossover}
\alias{CaseCrossover-package}
\title{CaseCrossover: Case-Crossover}
\description{
An R package for performing case-crossover and case-time-control analyses in an observational database in the OMOP Common Data Model.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://ohdsi.github.io/CaseCrossover}
  \item \url{https://github.com/OHDSI/CaseCrossover}
  \item Report bugs at \url{https://github.com/OHDSI/CaseCrossover/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
