% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseCrossoverData.R
\docType{class}
\name{CaseCrossoverData-class}
\alias{CaseCrossoverData-class}
\alias{CaseCrossoverData}
\alias{show,CaseCrossoverData-method}
\alias{summary,CaseCrossoverData-method}
\title{Case Data}
\usage{
\S4method{show}{CaseCrossoverData}(object)

\S4method{summary}{CaseCrossoverData}(object)
}
\arguments{
\item{object}{An object of class `CaseCrossoverData`.}
}
\description{
\code{CaseCrossoverData} is an S4 class that inherits from \code{\link[Andromeda]{Andromeda}}. It contains
information on cases.

A \code{CaseCrossoverData} object is typically created using \code{\link{getDbCaseCrossoverData}}, can only be saved using
\code{\link{saveCaseCrossoverData}}, and loaded using \code{\link{loadCaseCrossoverData}}.
}
\seealso{
\code{\link{isCaseCrossoverData}}
}
