% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{createCcrAnalysis}
\alias{createCcrAnalysis}
\title{Create a case-crossover analysis specification}
\usage{
createCcrAnalysis(analysisId = 1, description = "", exposureType = NULL,
  outcomeType = NULL, nestingCohortType = NULL, getDbCaseCrossoverDataArgs,
  selectSubjectsToIncludeArgs, getExposureStatusArgs)
}
\arguments{
\item{analysisId}{An integer that will be used later to refer to this specific set
of analysis choices.}

\item{description}{A short description of the analysis.}

\item{exposureType}{If more than one exposure is provided for each
exposureOutcomeNestingCohort, this field should be used to select
the specific exposure to use in this analysis.}

\item{outcomeType}{If more than one outcome is provided for each
exposureOutcomeNestingCohort, this field should be used to select
the specific outcome to use in this analysis.}

\item{nestingCohortType}{If more than one nesting cohort is provided for each
exposureOutcomeNestingCohort, this field should be used to select
the specific nesting cohort to use in this analysis.}

\item{getDbCaseCrossoverDataArgs}{An object representing the arguments to be used when calling the
\code{\link{createGetDbCaseCrossoverDataArgs}} function.}

\item{getExposureStatusArgs}{An object representing the arguments to be used when calling the
\code{\link{getExposureStatus}} function.}

\item{selectSubjectsToIncludesArgs}{An object representing the arguments to be used when calling the
\code{\link{selectSubjectsToInclude}} function.}
}
\description{
Create a case-crossover analysis specification
}
\details{
Create a set of analysis choices, to be used with the \code{\link{runCcrAnalyses}} function.
}
