% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createGetExposureStatusArgs}
\alias{createGetExposureStatusArgs}
\title{Create a parameter object for the function getExposureStatus}
\usage{
createGetExposureStatusArgs(firstExposureOnly = FALSE,
  riskWindowStart = -30, riskWindowEnd = 0, controlWindowOffsets = c(-60))
}
\arguments{
\item{firstExposureOnly}{Should only the first exposure per subject be included?}

\item{riskWindowStart}{The start of the risk window (in days) relative to the index date.This
number should be non-positive.}

\item{riskWindowEnd}{The end of the risk window (in days) relative to the index date.
Thisnumber should be non-positive.}

\item{controlWindowOffsets}{Offsets in days of the control windows relative to the case window.}
}
\description{
Create a parameter object for the function getExposureStatus
}
\details{
Create an object defining the parameter values.
}
