% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectSubjects.R
\name{createMatchingCriteria}
\alias{createMatchingCriteria}
\title{Create matching criteria}
\usage{
createMatchingCriteria(controlsPerCase = 2, matchOnAge = TRUE,
  ageCaliper = 2, matchOnGender = TRUE, matchOnProvider = FALSE,
  matchOnCareSite = FALSE, matchOnVisitDate = FALSE,
  visitDateCaliper = 30, matchOnTimeInCohort = FALSE,
  daysInCohortCaliper = 30)
}
\arguments{
\item{controlsPerCase}{Maximum number of controls to select per case.}

\item{matchOnAge}{Match on age?}

\item{ageCaliper}{Maximum difference (in years) in age when matching on age.}

\item{matchOnGender}{Match on gender?}

\item{matchOnProvider}{Match on provider (as specified in the person table)?}

\item{matchOnCareSite}{Match on care site (as specified in the person table)?}

\item{matchOnVisitDate}{Should the index date of the control be changed to the nearest visit
date?}

\item{visitDateCaliper}{Maximum difference (in days) between the index date and the visit
date when matching on visit date.}

\item{matchOnTimeInCohort}{Match on time in nesting cohort? When not using nesting, this is
interpreted as time observed prior to index.}

\item{daysInCohortCaliper}{Maximum difference (in days) in time in cohort.}
}
\description{
Create matching criteria
}
