% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFitting.R
\name{fitCaseCrossoverModel}
\alias{fitCaseCrossoverModel}
\title{Fit case-crossover model}
\usage{
fitCaseCrossoverModel(exposureStatus)
}
\arguments{
\item{exposureStatus}{A data frame as generated using the
\code{\link{getExposureStatus}} function.}
}
\description{
Fit case-crossover model
}
\details{
Fits a conditional logistic regression on the case-crossover data.
}
