% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-model-setup.R
\name{fortify_priors}
\alias{fortify_priors}
\title{Fortify a list of priors}
\usage{
fortify_priors(priorlist)
}
\arguments{
\item{priorlist}{A list of priors, each as created by e.g. pc_prior() or in the same format. Its
elements will be passed to validate_prior_distribution().}
}
\value{
A function which computes the joint log-prior of all hyperparameters, assuming prior
independence.
}
\description{
This function takes a list of priors, validates them using validate_prior_distribution(),
and stacks them together into one function that takes a vector argument, representing the joint prior
on all hyperparameters. This enforces prior independence between hyperparameters.

The user shouldn't have to
use this function; it is exported in case advanced users want to define their own priors.
}
