% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-likelihood.R
\name{get_casedays}
\alias{get_casedays}
\title{Get a vector of case day weights from the model_data}
\usage{
get_casedays(model_data)
}
\arguments{
\item{model_data}{A list of class "cc_modeldata" as returned by model_setup().}
}
\value{
A numeric vector of case day weights
}
\description{
Get a vector of case day weights, repeated once for every parameter
ascribed to each subject. Used inside gradient and hessian calculations as weights.
}
