% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-posthoc-quantities.R
\name{get_indices}
\alias{get_indices}
\title{Helper function to return the correct indices for latent variables}
\usage{
get_indices(model_data)
}
\arguments{
\item{model_data}{ccmodeldata object as returned by model_setup().}
}
\description{
The organization of the latent field is confusing. There
are random effects and fixed effects, plus the actual linear predictor values
(which you usually don't want).

This function accepts a ccmodeldata object and returns a list telling you
which indices in the vector of latent variables correspond to which model elements.

Details: the latent Gaussian variables have dimension Wd = Nd + M + p. Nd
is the total number of control days, sum(model_data$control_days), of all subjects.
M is the total number of random effects, i.e. unique values of smooth terms, in the
order in which they appear in the model. p is the number of regression coefficients.
To get the means/variances of the random effects, the correct call is
i = (model_data$Nd+1):(model_data$Nd+model_data$M). To get the regression coefficients,
the correct call is i = (model_data$Nd+model_data$M+1):(model_data$Nd+model_data$M+model_data$p).
}
