% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectSubjects.R
\name{getAttritionTable}
\alias{getAttritionTable}
\title{Get the attrition table for a set of subjects}
\usage{
getAttritionTable(subjects)
}
\arguments{
\item{subjects}{A data frame of subjects as generated by the function
\code{\link{selectSubjectsToInclude}}.}
}
\value{
A data frame specifying the number of cases and events after various steps of filtering.
}
\description{
Get the attrition table for a set of subjects
}
