% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-likelihood.R
\name{grad_log_likelihood}
\alias{grad_log_likelihood}
\title{Gadient of the case-crossover log-likelihood.}
\usage{
grad_log_likelihood(W, model_data)
}
\arguments{
\item{W}{Parameter vector. First n elements are eta, then Gamma and beta.}

\item{model_data}{A list of class "cc_modeldata" as returned by model_setup().}
}
\value{
A numeric vector containing the gradient. It's not stored as a sparseVector,
because it's (mathematically) not sparse. I guess the end is all zeroes though, so maybe
we should change this...?
}
\description{
Compute the gradient of the log likelihood with respect to W = (delta,gamma,beta).
The gamma and beta parts are zero.
}
