% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseCrossoverData.R
\name{isCaseCrossoverData}
\alias{isCaseCrossoverData}
\title{Check whether an object is a CaseCrossoverData object}
\usage{
isCaseCrossoverData(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A logical value.
}
\description{
Check whether an object is a CaseCrossoverData object
}
