% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{loadCcrAnalysisList}
\alias{loadCcrAnalysisList}
\title{Load a list of ccrAnalysis from file}
\usage{
loadCcrAnalysisList(file)
}
\arguments{
\item{file}{The name of the file}
}
\value{
A list of objects of type \code{ccrAnalysis}.
}
\description{
Load a list of objects of type \code{ccrAnalysis} from file. The file is in JSON format.
}
