% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Compute the log-likelihood for the model at a given parameter configuration}
\usage{
log_likelihood(W, model_data)
}
\arguments{
\item{W}{Parameter vector. First n elements are eta, then Gamma and beta.}

\item{model_data}{A list of class "cc_modeldata" as returned by model_setup().}
}
\value{
A number, the log-likelihood.
}
\description{
Compute the model log-likelihood for a given set of parameters.
}
