% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-latent-variables.R
\name{log_posterior_theta}
\alias{log_posterior_theta}
\alias{log_posterior_sigma}
\title{Log-posterior approximation for theta and sigma}
\usage{
log_posterior_theta(theta, W, model_data, hessian_structure = NULL, Q = NULL)

log_posterior_sigma(sigma, W, model_data, hessian_structure = NULL, Q = NULL)
}
\arguments{
\item{theta}{Vector of log-precisions at which to evaluate the log-posterior. theta = -2*log(sigma).}

\item{W}{The mode of the log-posterior of W|theta,y for this theta. Computed outside and passed in.}

\item{model_data}{a ccmodeldata object created using model_setup()}

\item{hessian_structure}{Optional, sparse structure of the Hessian created using hessian_log_likelihood_structure()}

\item{Q}{Optional, Q matrix evaluated at this theta. Will be created if not supplied.}

\item{sigma}{Vector of standard deviations at which to evaluate the log-posterior. sigma = exp(-.5 * theta)}
}
\description{
Compute the log-posterior for theta, the log-precision, and sigma, the standard deviation, of the RW2 model components.
}
