% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-posthoc-quantities.R
\name{make_linear_constraints}
\alias{make_linear_constraints}
\title{Make a matrix of linear constraints}
\usage{
make_linear_constraints(model_data)
}
\arguments{
\item{model_data}{ccmodeldata object created by model_setup. Must have control$linear_constraints}
}
\value{
A sparse matrix where each column represents one linear constraint. If the return value is A
then AW = 0.
}
\description{
Take the linear constraints from the control argument of a ccmodeldata object
and create a sparse matrix of linear constraints.
}
