% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-posthoc-quantities.R
\name{make_model_lincombs}
\alias{make_model_lincombs}
\title{Marginal means/variances for linear combinations of latent variables}
\usage{
make_model_lincombs(model_data)
}
\arguments{
\item{model_data}{ccmodeldata object output by model_setup()}
}
\value{
A sparse matrix with one column per necessary linear combination
}
\description{
The compute_marginal_means_and_variances() function lets you specify a sparse matrix of
linear combinations of latent variables to compute the marginal means and variances of. The most
common use case for this is when you have included an effect as both linear and smooth, and you
need the marginal mean and variance of the linear predictor. You need to account for the correlation
between the posterior regression coefficient and posterior random effect.

Because this is a common use case, this function takes in your model_data and returns a correctly
formatted matrix which specifies that you want means/variances for the linear predictors from all
terms which appear in the model both as linear and smooth effects.
}
