% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-posthoc-quantities.R
\name{marginal_hyperparameter_posterior}
\alias{marginal_hyperparameter_posterior}
\title{Marginal hyperparameter posterior}
\usage{
marginal_hyperparameter_posterior(j, ccfit, quantiles = c(2.5, 97.5)/100)
}
\arguments{
\item{j}{Integer index indicating which marginal to evaluate. If theta is K-dimensional then this should be from 1 to K.}

\item{ccfit}{Object of class ccfit returned by casecrossover()}

\item{quantiles}{Vector of numbers between 0 and 1 representing quantiles of the marginal posterior that are to be estimated}
}
\value{
A list containing elements
        - margpost: dataframe of logged theta_j posterior values, evaluated on a grid of the same accuracy as those used
to construct the full joint.
        - quantiles: dataframe containing the quantiles of the marginal posterior. Easier to do these here since they
        require a grid.
}
\description{
Get the marginal density of a single coordinate of the hyperparameter vector theta using
quadrature.
}
