% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-posthoc-quantities.R
\name{normalize_optresults_logpost}
\alias{normalize_optresults_logpost}
\title{Normalize the log posterior returned by the optimization}
\usage{
normalize_optresults_logpost(optresults)
}
\arguments{
\item{optresults}{Optimization results, a tibble() output by optimize_all_thetas_parallel().}
}
\value{
Data frame of optimization results with the theta_logposterior column normalized,
i.e. satisfying logSumExp(theta_logposterior) = 0
}
\description{
Wrapper around normalize_log_posterior() that takes in
the dataframe of optimization results and returns the same dataframe
but with the theta_logposterior column normalized.
}
