% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-optimization.R
\name{optimize_latentfield_trustoptim}
\alias{optimize_latentfield_trustoptim}
\alias{optimize_all_thetas_parallel}
\title{Optimize the conditional posterior of the latent Gaussian variables for one hyperparameter configuration}
\usage{
optimize_latentfield_trustoptim(
  theta,
  model_data,
  hessian_structure = NULL,
  Q = NULL,
  optcontrol = NULL
)

optimize_all_thetas_parallel(
  thetagrid,
  model_data,
  hessian_structure = NULL,
  optcontrol = NULL,
  doparallel = TRUE
)
}
\arguments{
\item{theta}{Vector, representing one configuration of hyperparameters (optimize_latentfield_trustoptim). List of vectors, each
representing one configuration of hyperparameters (optimize_all_thetas_parallel).}

\item{model_data}{ccmodeldata object as output by model_setup().}

\item{hessian_structure}{Optional, provide the sparsity structure of the Hessian. This doesn't change with different thetas, so when
doing all hyperparameter configurations, only need to compute this once. It will be computed if not provided.}

\item{Q}{Optional, provide pre-computed Q matrix for this theta. Will be computed if not provided.}

\item{optcontrol}{Optional. Override the casecrossover default control parameters for trust.optim(). See cc_control()$opt_control.}

\item{thetagrid}{A NIGrid object which contains the grid of theta values to be optimized for.}

\item{doparallel}{Logical. Use parallel::mclapply (TRUE) or lapply (FALSE) for execution? Former executes in parallel, latter in series. Default TRUE.}
}
\description{
Compute the conditional mode W-hat for one value of the hyperparameters theta. Uses
trust region methods implemented in trustOptim::trust.optim. Makes use of the sparsity of the Hessian.

This is a wrapper for optimize_latentfield_trustoptim() that computes the
conditional mode in parallel over a user-specified grid of theta values.

Parallelization only works on linux and mac; the "parallel" package needs to work. If you're on windows, you'll just
have to wait a little longer (long enough to go to the store and buy a serious computer maybe?).
}
