% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-optimization.R
\name{optlist_to_tibble}
\alias{optlist_to_tibble}
\title{Convert the output of one of my optimization functions into a tibble for easy analysis.}
\usage{
optlist_to_tibble(lst)
}
\arguments{
\item{lst}{A list formatted according to the output of optimize_latentfield_trustoptim().}
}
\description{
Not exported; for development use only.
}
