% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-model-setup.R
\name{pc_prior}
\alias{pc_prior}
\title{The PC Prior for the smoothing log precision.}
\usage{
pc_prior(u, alpha)
}
\arguments{
\item{u}{The "u" in P(sigma > u) = 1 - alpha.}

\item{alpha}{The "alpha" in P(sigma > u) = 1 - alpha.}
}
\description{
The PC prior for the smoothing log precision is described in the package vignette.
This is a convenience function for specifying this prior in the casecrossover software.

The prior is specified as P(sigma > u) = 1 - alpha.
}
\examples{
pc_prior(u = 3,alpha = .5)

}
