% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-casecrossover.R
\name{plot.cc_fit}
\alias{plot.cc_fit}
\title{Plot results of fitting a case crossover model.}
\usage{
\method{plot}{cc_fit}(x, ...)
}
\arguments{
\item{x}{Object of class cc_fit returned by casecrossover().}

\item{...}{Not used.}
}
\value{
A named list of lists of ggplots, one for the fixed effects and one for the smooth.
Only mimimal annotations are made; you can put your own titles and axis labels and further change the plot
features using the ggplot2 API. The plot lists themselves can be passed directly to cowplot::plot_grid,
or further annotated by you.
}
\description{
Plots corresponding to a case crossover model. This function computes histograms of
fixed-effect posteriors and smooth curves with pointwise error bars (+- 2sd) based on the marginal
posteriors of the smooth effects.
}
