% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-casecrossover.R
\name{print.cc_summary}
\alias{print.cc_summary}
\title{Print method for summary.cc_fit.}
\usage{
print.cc_summary(ccsummary)
}
\arguments{
\item{ccsummary}{An object of class cc_summary, output by summary().}
}
\value{
Nothing.
}
\description{
Generic print method for a cc_fit summary object. Not exported, user will call this function
by simply typing summary(...) in the console, like normal.
}
