% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-latent-variables.R
\name{Q_matrix_linear}
\alias{Q_matrix_linear}
\alias{Q_matrix_rw2_one_component}
\alias{Q_matrix_rw2}
\alias{Q_matrix_both}
\alias{Q_matrix}
\title{Calculate the Q matrix for a model with only linear terms.}
\usage{
Q_matrix_linear(model_data, tau = exp(12))

Q_matrix_rw2_one_component(theta, model_data, covariate)

Q_matrix_rw2(theta, model_data, tau = exp(12))

Q_matrix_both(theta, model_data, tau = exp(12))

Q_matrix(theta, model_data, tau = exp(12))
}
\arguments{
\item{model_data}{A cc_modeldata object returned by model_setup()}

\item{tau}{The precision of the auxillary epsilon variables. Set it to something huge. Default exp(12).}

\item{theta}{Hyperparameter, scalar, the log-precision of the RW2 model.}

\item{covariate}{String naming the covariate from the data for which the precision matrix should
be calculated.}
}
\value{
A sparse matrix inheriting from class CsparseMatrix containing the joint precision matrix of the
latent Gaussian variables.
}
\description{
Calculate the Q-matrix (joint prior precision matrix) of the latent Gaussian variables.
This is a function of theta. Functions are provided for Q-matrices with linear terms only,
RW2 terms only, and both.

Note that the linear terms' prior covariance matrix is fixed (a true prior) and its log-precision
is specified directly in the model setup. Hence the Q_matrix_linear function does not take an argument
for this parameter.
}
