% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{runCcrAnalyses}
\alias{runCcrAnalyses}
\title{Run a list of analyses}
\usage{
runCcrAnalyses(
  connectionDetails,
  cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema,
  exposureDatabaseSchema = cdmDatabaseSchema,
  exposureTable = "drug_era",
  outcomeDatabaseSchema = cdmDatabaseSchema,
  outcomeTable = "condition_era",
  nestingCohortDatabaseSchema = cdmDatabaseSchema,
  nestingCohortTable = "condition_era",
  outputFolder = "./CcrOutput",
  ccrAnalysisList,
  exposureOutcomeNestingCohortList,
  getDbCaseCrossoverDataThreads = 1,
  selectSubjectsToIncludeThreads = 1,
  getExposureStatusThreads = 1,
  fitCaseCrossoverModelThreads = 1
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{ConnectionDetails} created using
the function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On
SQL Server, this should specify both the database and the
schema, so for example 'cdm_instance.dbo'.}

\item{oracleTempSchema}{A schema where temp tables can be created in Oracle.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where
the exposure data used to define the exposure cohorts is
available. If exposureTable = DRUG_ERA,
exposureDatabaseSchema is not used but assumed to be
cdmSchema.  Requires read permissions to this database.}

\item{exposureTable}{The tablename that contains the exposure cohorts.  If
exposureTable <> drug_era, then expectation is
exposureTable has format of COHORT table:
cohort_definition_id, subject_id, cohort_start_date,
cohort_end_date.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where
the data used to define the outcome cohorts is available.
If outcomeTable = CONDITION_ERA, outcomeDatabaseSchema is
not used.  Requires read permissions to this database.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  If
outcomeTable is not CONDITION_OCCURRENCE or CONDITION_ERA,
then expectation is outcomeTable has format of COHORT
table: COHORT_DEFINITION_ID, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{nestingCohortDatabaseSchema}{The name of the database schema that is the location where
the nesting cohort is defined.}

\item{nestingCohortTable}{Name of the table holding the nesting cohort. This table
should have the same structure as the cohort table.}

\item{outputFolder}{Name of the folder where all the outputs will written to.}

\item{ccrAnalysisList}{A list of objects of type \code{ccrAnalysis} as created
using the \code{\link{createCcrAnalysis}} function.}

\item{exposureOutcomeNestingCohortList}{A list of objects of type
\code{exposureOutcomeNestingCohort} as created using the
\code{\link{createExposureOutcomeNestingCohort}} function.}

\item{getDbCaseCrossoverDataThreads}{The number of parallel threads to use for building the
caseControlData objects.}

\item{selectSubjectsToIncludeThreads}{The number of parallel threads to use for selecting
subjects}

\item{getExposureStatusThreads}{The number of parallel threads to use for getting exposure
status.}

\item{fitCaseCrossoverModelThreads}{The number of parallel threads to use for fitting the
models.}
}
\description{
Run a list of analyses
}
\details{
Run a list of analyses for the exposure-outcome-nesting cohorts of interest. This function will run
all specified analyses against all hypotheses of interest, meaning that the total number of outcome
models is `length(ccrAnalysisList) * length(exposureOutcomeNestingCohortList)`. When you provide
several analyses it will determine whether any of the analyses have anything in common, and will
take advantage of this fact. For example, if we specify several analyses that only differ in the way
the control windows are specified then this function will extract the data and select the subjects
only once, and re-use this in all the analysis.
}
