% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseCrossoverData.R
\name{saveCaseCrossoverData}
\alias{saveCaseCrossoverData}
\title{Save the case data to file}
\usage{
saveCaseCrossoverData(caseCrossoverData, file)
}
\arguments{
\item{caseCrossoverData}{An object of type \code{CaseCrossoverData} as generated using
\code{\link{getDbCaseCrossoverData}}.}

\item{file}{The name of the file where the data will be written. If the file
exists it will be overwritten.}
}
\description{
\code{saveCaseCrossoverData} saves an object of type caseCrossoverData to file
}
\details{
The data will be written to the file specified by the user.
}
