% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{saveCcrAnalysisList}
\alias{saveCcrAnalysisList}
\title{Save a list of ccrAnalysis to file}
\usage{
saveCcrAnalysisList(ccrAnalysisList, file)
}
\arguments{
\item{ccrAnalysisList}{The ccArnalysis list to be written to file}

\item{file}{The name of the file where the results will be written}
}
\description{
Write a list of objects of type \code{ccrAnalysis} to file. The file is in JSON format.
}
