% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-casecrossover.R
\name{summary.cc_fit}
\alias{summary.cc_fit}
\title{Summarize results of a casecrossover fit.}
\usage{
\method{summary}{cc_fit}(object, ...)
}
\arguments{
\item{object}{Object of class cc_fit returned by casecrossover().}

\item{...}{Not used}
}
\value{
A list of summary information of class cc_summary, with a print method.
}
\description{
Generic summary method for a cc_fit object returned by casecrossover(). Prints similar
information to the familiar lm() and glm() summary functions, or inla().
}
