% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-model-setup.R
\name{supported_prior_distributions}
\alias{supported_prior_distributions}
\title{Supported prior distributions.}
\usage{
supported_prior_distributions(nm = "")
}
\arguments{
\item{nm}{Optional; if you know the name of the prior you want, specify it.}
}
\description{
This function returns a list of all supported prior distributions,
including their names, parameters, and actual function call.
}
\examples{
supported_prior_distributions()
supported_prior_distributions("pc.prec")

}
