% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-model-setup.R
\name{validate_linear_constraints}
\alias{validate_linear_constraints}
\title{Validate your created linear constraints.}
\usage{
validate_linear_constraints(constraints, verbose = TRUE)
}
\arguments{
\item{constraints}{What you're planning on inputting to cc_control(linear_constraints = ...)}

\item{verbose}{Logical. Should debugging information be printed?}
}
\description{
This function checks the format of what you're inputting into cc_control() for
linear constraints. This is to help you format your model correctly and prevent downstream
errors, which can be harder to debug.
}
\examples{
validate_linear_constraints(create_linear_constraints(u = c(1,2,3),whichzero = 1)) # Valid
validate_linear_constraints(list(x = c(1,2,3))) # Not valid

}
