% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-model-setup.R
\name{validate_prior_distribution}
\alias{validate_prior_distribution}
\title{Validate a specified prior distribution.}
\usage{
validate_prior_distribution(control)
}
\arguments{
\item{control}{A control list as returned by cc_control().}
}
\description{
Pass in your control parameters, and this function will tell you whether you specified
your priors correctly. It just checks supported_prior_distributions() for you; so also, see that function.
You won't get a formal error if your prior is determined to be incorrectly specified; you'll get a
message that tries to help you fix it.
}
\examples{
mycontrol <- cc_control(smooth_prior = list(name = "pc.prec",params = c(u = 3,alpha = .5)))
mycontrol_missinginfo <- cc_control(smooth_prior = list(name = "pc.prec"))
mycontrol_badprior <- cc_control(smooth_prior = list(name = "foo"))

validate_prior_distribution(mycontrol)
validate_prior_distribution(mycontrol_missinginfo)
validate_prior_distribution(mycontrol_badprior)

}
