% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArgFunctions.R
\name{createGetDbCaseCrossoverDataArgs}
\alias{createGetDbCaseCrossoverDataArgs}
\title{Create a parameter object for the function getDbCaseCrossoverData}
\usage{
createGetDbCaseCrossoverDataArgs(useNestingCohort = FALSE,
  useObservationEndAsNestingEndDate = TRUE, getVisits = FALSE,
  studyStartDate = "", studyEndDate = "", getTimeControlData = FALSE,
  maxNestingCohortSize = 1e+07, maxCasesPerOutcome = 5e+05)
}
\arguments{
\item{useNestingCohort}{Should the study be nested in a cohort (e.g. people witha
specific indication)? If not, the study will be nestedin
the general population.}

\item{useObservationEndAsNestingEndDate}{When using a nesting cohort, should the observationperiod
end date be used instead of the cohort end date?}

\item{getVisits}{Get data on visits? This is needed when performing atime-
case-control study and matching on visit date isrequested
later on.}

\item{studyStartDate}{A calendar date specifying the minimum date where data
isused. Date format is 'yyyymmdd'.}

\item{studyEndDate}{A calendar date specifying the maximum date where data
isused. Date format is 'yyyymmdd'.}

\item{getTimeControlData}{Should data for time controls be fetched? (needed
forcase-time-control analyses).}

\item{maxNestingCohortSize}{If the nesting cohort is larger thanthis number it will
be sampled to this size. maxCohortSize = 0indicates no
maximum size. (needed for case-time-control analyses).}

\item{maxCasesPerOutcome}{If there are more than this number of cases for a
singleoutcome cases will be sampled to this size.
maxCasesPerOutcome = 0indicates no maximum size.}
}
\description{
Create a parameter object for the function getDbCaseCrossoverData
}
\details{
Create an object defining the parameter values.
}
