% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseCrossoverData.R
\name{loadCaseCrossoverData}
\alias{loadCaseCrossoverData}
\title{Load the covariate data from a folder}
\usage{
loadCaseCrossoverData(file)
}
\arguments{
\item{file}{The name of the file containing the data.}
}
\value{
An object of class \code{CaseCrossoverData}.
}
\description{
\code{loadCaseCrossoverData} loads an object of type caseCrossoverData from a file in the file system.
}
\details{
The data will be read from the file specified by the user.
}
