% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-posthoc-quantities.R
\name{normalize_log_posterior}
\alias{normalize_log_posterior}
\title{Normalize log-posterior via simple numerical integration}
\usage{
normalize_log_posterior(pp, tt)
}
\arguments{
\item{pp}{logged function values}

\item{tt}{grid points at which these function values were computed, as a NIGrid object}
}
\value{
A number giving the LOG of the integral of the function used to compute pp.
This is the log-normalizing constant.
}
\description{
Simple implementation of the trapezoid rule for integration. The key is that
this function accepts grid points and LOGGED function values and computes the integral in
a numerically stable manner, using matrixStats::logSumExp().

For a single hyperparameter, the function works on uniform and non-uniform grids. For
a multidimensional hyperparameter, the function only works on a uniform grid (in all dimensions).
}
