% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-casecrossover.R
\name{print.cc_summary}
\alias{print.cc_summary}
\title{Print method for summary.cc_fit.}
\usage{
print.cc_summary(x)
}
\arguments{
\item{x}{An object of class cc_summary, output by applying summary() to an object of class cc_fit returned by casecrossover().}
}
\value{
Nothing.
}
\description{
Generic print method for a cc_fit summary object. Not exported, user will call this function
by simply typing summary(...) in the console, like normal.
}
