% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadingSaving.R
\name{saveCaseCrossoverData}
\alias{saveCaseCrossoverData}
\title{Save the case-crossover data to folder}
\usage{
saveCaseCrossoverData(caseCrossoverData, folder)
}
\arguments{
\item{folder}{The name of the folder where the data will be written. The folder should not yet
exist.}

\item{caseCrossover}{An object of type \code{caseCrossoverData} as generated using \code{\link{getDbCaseCrossoverData}}.}
}
\description{
\code{saveCaseCrossoverData} saves an object of type caseCrossoverData to folder.
}
\details{
The data will be written to a set of files in the specified folder.
}
