% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnalyses.R
\name{summarizeCcrAnalyses}
\alias{summarizeCcrAnalyses}
\title{Create a summary report of the analyses}
\usage{
summarizeCcrAnalyses(outcomeReference)
}
\arguments{
\item{outcomeReference}{A data.frame as created by the \code{\link{runCcrAnalyses}} function.}
}
\description{
Create a summary report of the analyses
}
